[edit] consolidating, moving and renaming files
This the beginning of a major refactoring and reorganization of w&w.

The new organization will break the work into 4 volumes (it remains to
be seen if that will also be four physical volumes or not): Its, Bits,
Domas and Polis.  Everything else be written up as standard stand
alone patterns in the same mold as APL.